Remove the only dependence to file buffer variable
File buffer variable access through the caret syntax (input^) is not
supported in FPC and Delphi.

This is used in one place: when recognizing a number and encountering a
point, it tests if the point is followed by another point (i.e. if it's
a decimal point or the '..' operator).

In fact, the code from the print version of the book doesn't depend on
buffer accesses, so I reverted back this part to the original code (the
differences between the printed and online version are available at
http://homepages.cwi.nl/~steven/pascal/).

But now, we have a goto which point inside an else clause but FPC and
Delphi support that.