Replace Read(Integer) by ISO-like equivalent
In ISO Pascal, only the characters forming the number are read whereas
in Delphi and Free Pascal, the characters are read until the next blank.

This means that with '123=456', Read(Int) would return 123 in ISO and
the next char to be read would be '='. And would generate an error with
a Borland-like compiler as '123=' is not a number.

The ISO behaviour is needed in two places:
  - when reading a label as it can be followed directly by '='
  - when reading a set literal as it can be followed by ')'