Fix arch dependency with ORD and CHR instr.
ORD and CHR are NOOP in this implementation as it is assumed that
integers, chars and booleans are interchangeable.

CHR (opcode 60) is now assembled without operands and its
implementation copies the .vc field into .vi.

ORD (opcode 59) is assembled with a type operand. ORDR can be generated
by the compiler but I won't implement it for now. The implementation
copies the corresponding field to .vi.

The assembled code is longer now as CHR and ORD are assembled so I had
to increase the code store to be able to run pcom.

An advantage to this is that the instruction numbering (i labels) the
compiler generates are now accurate.

This modification is based on the work of Scott Moore.