Fix arch dependency in DEC implementation
As for DEC, it was assumed that integers, chars, booleans and even
reals were interchangeable, so the DEC impl constisted in decrementing
the .vi field.

Also, removed DEC on set as this instruction doesn't exist and DEC on
adress as this instruction is never generated by the compiler.

Based on the work of Scott Moore.