Fix number alignment problems in code generator
When generating LDC(, numbers are aligned on 3 chars. A set member >99
would not be padded with spaces and so would extend the previous number.

Simarly in gen2t which is notably used by CHKA and CHKI the numbers are
aligned on 8 chars. This is not enough if maxint > 99 999 999.

By example:
 ldc( 100101)
 chka 12147483647